Moving neccessery functions to tools.py
