+++ demisto_sdk/commands/common/tools.py
@@ -20,7 +20,7 @@ from demisto_sdk.commands.common.constants import (
     DEF_DOCKER, DEF_DOCKER_PWSH, INTEGRATIONS_DIR, LAYOUTS_DIR,
     PACKAGE_SUPPORTING_DIRECTORIES, PACKAGE_YML_FILE_REGEX, PACKS_DIR,
     PACKS_DIR_REGEX, PACKS_README_FILE_NAME, RELEASE_NOTES_REGEX, SCRIPTS_DIR,
-    SDK_API_GITHUB_RELEASES, TESTS_DIRECTORIES, TYPE_PWSH, UNRELEASE_HEADER)
+    SDK_API_GITHUB_RELEASES, TESTS_DIRECTORIES, TYPE_PWSH, UNRELEASE_HEADER, CONF_PATH)
 
 # disable insecure warnings
 urllib3.disable_warnings()
@@ -45,7 +45,7 @@ def get_log_verbose() -> bool:
     return LOG_VERBOSE
 
 
-def get_yml_paths_in_dir(project_dir: str, error_msg: str = '',) -> Tuple[list, str]:
+def get_yml_paths_in_dir(project_dir: str, error_msg: str = '', ) -> Tuple[list, str]:
     """
     Gets the project directory and returns the path of the first yml file in that directory
     :param project_dir: string path to the project_dir
@@ -801,3 +801,62 @@ def get_depth(data: Any) -> int:
     if data and isinstance(data, list):
         return 1 + max(get_depth(element) for element in data)
     return 0
+
+
+def find_test_match(test_config: dict, test_playbook_id: str, content_item_id: str, file_type: str) -> bool:
+    """
+    Given a test configuration from conf.json file, this method checks if the configuration is configured for the
+    test playbook with content item.
+    Since in conf.json there could be test configurations with 'integrations' as strings or list of strings
+    the type of test_configurations['integrations'] is checked in first and the match according to the type.
+    If file type is not an integration- will return True if the test_playbook id matches playbookID.
+    Args:
+        file_type: The file type. can be 'integration', 'script', 'playbook'.
+        test_config: A test configuration from conf.json file under 'tests' key.
+        test_playbook_id: A test playbook ID.
+        content_item_id: A content item ID, could be a script, an integration or a playbook.
+
+    Returns:
+        True if the test configuration contains the test playbook and the content item or False if not
+    """
+    if test_playbook_id != test_config.get('playbookID'):
+        return False
+    if file_type != 'integration':
+        return True
+
+    test_integrations = test_config.get('integrations')
+    if isinstance(test_integrations, list):
+        return any(
+            test_integration for test_integration in test_integrations if test_integration == content_item_id)
+    else:
+        return test_integrations == content_item_id
+
+
+def get_not_registered_tests(conf_json_tests: list, content_item_id: str, file_type: str, test_playbooks: list) -> list:
+    """
+    Return all test playbooks that are not configured in conf.json file
+    Args:
+        conf_json_tests: the 'tests' value of 'conf.json file
+        content_item_id: A content item ID, could be a script, an integration or a playbook.
+        file_type: The file type, could be a script, an integration or a playbook.
+        test_playbooks: The yml file's list of test playbooks
+
+    Returns:
+        A list of TestPlaybooks not configured
+    """
+    not_registered_tests = []
+    for test in test_playbooks:
+        test_registered_in_conf_json = any(
+            test_config for test_config in conf_json_tests if find_test_match(test_config,
+                                                                              test,
+                                                                              content_item_id,
+                                                                              file_type)
+        )
+        if not test_registered_in_conf_json:
+            not_registered_tests.append(test)
+    return not_registered_tests
+
+
+def load_conf_file():
+    with open(CONF_PATH) as data_file:
+        return json.load(data_file)
